# lib/api-zod/src/generated/types/aiVideoRenderBody.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Homework auto-grading platform API
 * OpenAPI spec version: 0.1.0
 */

export interface AiVideoRenderBody {
  /**
     * @minLength 8
     * @maxLength 100
     */
  idempotencyKey: string;
}