# lib/api-zod/src/generated/types/aiVideoStoryboard.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Homework auto-grading platform API
 * OpenAPI spec version: 0.1.0
 */
import type { AiVideoScene } from './aiVideoScene';

export interface AiVideoStoryboard {
  /**
     * @minLength 1
     * @maxLength 160
     */
  title: string;
  /** @minimum 1 */
  version: number;
  /**
     * @minItems 5
     * @maxItems 10
     */
  scenes: AiVideoScene[];
}